Camera Follow algorithm with constraints. Battlefield fixed.
The orthographic size for the camera is calculated based on the stop distance and the width of the focus area. In addition, the maximum movement size of the focus area is calculated so that the camera stops accurately at the boundaries of the battlefield if the camera margin value is set to zero.
The calculations take into account the current aspect of the camera, as well as the look ahead distance.